requirements.txt
pyinstaller
python-dotenv
Pillow
qrcode
pyserial
pandas
gspread
oauth2client
